Fix errors in image and label loading

+++ include/core/common_types.h
@@ -5,7 +5,7 @@
 #include <vector>
 
 namespace neural_net {
-typedef std::vector<std::vector<std::vector<double>>> Weight_Collection_t;
-typedef std::vector<std::vector<double>> Image_t;
+typedef std::vector<std::vector<std::vector<float>>> Weight_Collection_t;
+typedef std::vector<std::vector<float>> Image_t;
 
 } // namespace neural_net
