+++ include/core/neuron.h
@@ -18,23 +18,23 @@ class Neuron {
    * @param values the activation values from the previous layer
    * @return the calculated activation value of the neuron
    */
-  double ForwardPass(const std::vector<double>& weights, const std::vector<double>& values);
+  float ForwardPass(const std::vector<float>& weights, const std::vector<float>& values);
 
-  double GetActivation();
+  float GetActivation();
 
-  void CalculateError(double prev_error);
+  void CalculateError(float prev_error);
 
-  double GetError();
+  float GetError();
 
-  void SetError(double error);
+  void SetError(float error);
 
-  void IncrementDelta(double next_error);
+  void IncrementDelta(float next_error);
 
   void CalculateGradient(size_t batch_size);
 
-  double GetGradient();
+  float GetGradient();
 
-  void SetActivation(double activation);
+  void SetActivation(float activation);
 
  private:
   /**
@@ -42,10 +42,10 @@ class Neuron {
    * @param value value to be mapped
    * @return the adjusted value
    */
-  double Sigmoid(double value);
-  double activation_ = 0;
-  double error_;
-  double delta_;
-  double gradient_;
+  float Sigmoid(float value);
+  float activation_ = 0;
+  float error_;
+  float delta_;
+  float gradient_;
 };
 } // namespace neural_net
